# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: HRM + Distributor ERP Platform API — Phase 1 (Auth, Employees, Attendance, Leave)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/companies:
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company (initial tenant setup)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /v1/companies/me:
    get:
      operationId: getMyCompany
      tags: [companies]
      summary: Get current tenant company details
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateMyCompany
      tags: [companies]
      summary: Update company settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ── Departments ────────────────────────────────────────────────────────
  /v1/departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments in the company
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Department list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a new department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /v1/departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Department details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # ── Job Titles ─────────────────────────────────────────────────────────
  /v1/job-titles:
    get:
      operationId: listJobTitles
      tags: [job-titles]
      summary: List all job titles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Job title list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobTitle"
    post:
      operationId: createJobTitle
      tags: [job-titles]
      summary: Create a job title
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobTitleInput"
      responses:
        "201":
          description: Job title created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobTitle"
  /v1/job-titles/{id}:
    patch:
      operationId: updateJobTitle
      tags: [job-titles]
      summary: Update a job title
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobTitleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobTitle"
    delete:
      operationId: deleteJobTitle
      tags: [job-titles]
      summary: Delete a job title
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # ── Employees ──────────────────────────────────────────────────────────
  /v1/employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees (paginated, filterable)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: departmentId
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [ACTIVE, ON_LEAVE, TERMINATED, RESIGNED, SUSPENDED, null]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: employmentType
          in: query
          schema:
            type: ["string", "null"]
            enum: [FULL_TIME, PART_TIME, CONTRACT, INTERN, null]
      responses:
        "200":
          description: Paginated employee list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
  /v1/employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: terminateEmployee
      tags: [employees]
      summary: Terminate employee (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # ── Attendance ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
      required: [success, message, timestamp]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ── Auth Schemas ────────────────────────────────────────────────────────
...
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        taxId:
          type: ["string", "null"]
        currency:
          type: string
        timezone:
          type: string
        fiscalYearStart:
          type: integer
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, currency, timezone, fiscalYearStart, createdAt, updatedAt]
    CompanyInput:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        taxId:
          type: ["string", "null"]
        currency:
          type: string
          default: "INR"
        timezone:
          type: string
          default: "Asia/Kolkata"
        fiscalYearStart:
          type: integer
          default: 4
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
      required: [name]
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        taxId:
          type: ["string", "null"]
        currency:
          type: string
        timezone:
          type: string
        fiscalYearStart:
          type: integer
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]

    # ── Department Schemas ──────────────────────────────────────────────────
    Department:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        code:
          type: string
        parentId:
          type: ["string", "null"]
        headId:
          type: ["string", "null"]
        headName:
          type: ["string", "null"]
        employeeCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyId, name, code, createdAt, updatedAt]
    DepartmentInput:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        parentId:
          type: ["string", "null"]
        headId:
          type: ["string", "null"]
      required: [name, code]
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        parentId:
          type: ["string", "null"]
        headId:
          type: ["string", "null"]

    # ── Job Title Schemas ───────────────────────────────────────────────────
    JobTitle:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        level:
          type: integer
        minSalary:
          type: ["number", "null"]
        maxSalary:
          type: ["number", "null"]
        createdAt:
          type: string
      required: [id, title, level, createdAt]
    JobTitleInput:
      type: object
      properties:
        title:
          type: string
        level:
          type: integer
        minSalary:
          type: ["number", "null"]
        maxSalary:
          type: ["number", "null"]
      required: [title, level]
    JobTitleUpdate:
      type: object
      properties:
        title:
          type: string
        level:
          type: integer
        minSalary:
          type: ["number", "null"]
        maxSalary:
          type: ["number", "null"]

    # ── Employee Schemas ────────────────────────────────────────────────────
    EmployeeSummary:
      type: object
      properties:
        id:
          type: string
        employeeCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        department:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
          required: [id, name]
        jobTitle:
          type: object
          properties:
            id:
              type: string
            title:
              type: string
          required: [id, title]
        status:
          type: string
        employmentType:
          type: string
        hireDate:
          type: string
          format: date
      required: [id, employeeCode, firstName, lastName, email, status, employmentType, hireDate]
    Employee:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        employeeCode:
          type: string
        companyId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        department:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
          required: [id, name]
        jobTitle:
          type: object
          properties:
            id:
              type: string
            title:
              type: string
            level:
              type: integer
          required: [id, title, level]
        manager:
          $ref: "#/components/schemas/ManagerRef"
        status:
          type: string
        employmentType:
          type: string
        hireDate:
          type: string
          format: date
        probationEndDate:
          type: ["string", "null"]
        terminationDate:
          type: ["string", "null"]
        baseSalary:
          type: number
        currency:
          type: string
        address:
          $ref: "#/components/schemas/AddressField"
        emergencyContact:
          $ref: "#/components/schemas/EmergencyContactField"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, employeeCode, companyId, firstName, lastName, email, status, employmentType, hireDate, baseSalary, currency, createdAt, updatedAt]
    EmployeeListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSummary"
        pagination:
          $ref: "#/components/schemas/Pagination"
        timestamp:
          type: string
      required: [success, data, pagination, timestamp]
    EmployeeInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          default: "EMPLOYEE"
        phone:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        departmentId:
          type: string
        jobTitleId:
          type: string
        managerId:
          type: ["string", "null"]
        employmentType:
          type: string
          default: FULL_TIME
        hireDate:
          type: string
          format: date
        probationEndDate:
          type: ["string", "null"]
        baseSalary:
          type: number
        currency:
          type: string
          default: INR
        address:
          $ref: "#/components/schemas/AddressField"
        emergencyContact:
          $ref: "#/components/schemas/EmergencyContactField"
      required: [firstName, lastName, email, password, departmentId, jobTitleId, hireDate, baseSalary]
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        departmentId:
          type: string
        jobTitleId:
          type: string
        managerId:
          type: ["string", "null"]
        employmentType:
          type: string
        status:
          type: string
        probationEndDate:
          type: ["string", "null"]
        baseSalary:
          type: number
        address:
          $ref: "#/components/schemas/AddressField"
        emergencyContact:
          $ref: "#/components/schemas/EmergencyContactField"
...
    ManagerRef:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        photoUrl:
          type: string
      required: [id, firstName, lastName]
      additionalProperties: false
    AddressField:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
      additionalProperties: false
    EmergencyContactField:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        relation:
          type: string
      additionalProperties: false