# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HRM + Distributor ERP Platform API — Phase 1 (Auth, Employees, Attendance, Leave)
 * OpenAPI spec version: 1.0.0
 */

export interface Company {
  id: string;
  name: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  taxId?: string | null;
  currency: string;
  timezone: string;
  fiscalYearStart: number;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  createdAt: string;
  updatedAt: string;
}